requirements.txt
slack-bolt==1.18.1
google-generativeai==0.7.2
supabase==2.5.0
python-dotenv==1.0.1
